postgres_ffi: move `WalGenerator` into separate module
